booking-agent/pyproject.toml
[project]
name = "booking-agent"
version = "0.1.0"
description = "Booking Assistant Agent using Microsoft Agent Framework"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agent-framework-azure-ai>=0.1.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]
